Complete 2020 Day 16 Part 1, and most of Part 2
So far, my part 2 solution is able to determine all but 2 of the correct positions for the fields, but is unable to find any valid options for the 2 remaining fields. To get the second star, I found what the products would be under both options for the field orders, and tried to submit both to the website (with the second attempt working). I plan to try to make part 2 work fully later today.