Complete 2021 Day 24
Today's puzzle was only completable because I read a lot of tips and tricks on the subreddit, so thank you to everybody over there!

Without external help, I managed to work out that there was a consistent set of 14 instructions that would run for each digit in the input, and I could see that there were three values that differed for each digit. However, it wasn't until reading through the subreddit that I realised:
• The z variable is implementing a stack in base 26
• The values changing for each digit are implementing constrains between pairs of digits.

Once I realised this last point, I managed to work out how the pairs were constrained and implemented some logic to generate these (a 7-deep nested for loop over specific ranges). Unfortunately, this does mean that (in this current implementation) the solutions generated will only ever work for my input (and will thus fail the check against the ALU at the end), but it would be possible to work out these constraints programmatically and thus generate the ranges automatically. Maybe another day :sleeping: