Complete 2023 Day 24 Part 2
This was written over the course of the afternoon/evening of 24th December, but I've only committed it now as that's when I've been able to submit my solution. On my PC, it took 174811.976 seconds (48 hours, 33 minutes, and 31.976 seconds). Notably, it is also wrong. On my input, it returned an answer one fewer than the actual answer (with a solution found using z3 and which will be committed after this version).

I've added functionality to prevent a solution from being run if on a GitHub Actions runner and enabled it for this part. Since I've now re-written it using z3, it's a bit redundant to have included, but may come in useful in the future.